Remove projects fixes #458
