+++ app/experimenter/experiments/tests/test_api_views.py
@@ -11,7 +11,6 @@ from experimenter.experiments.serializers import (
     ExperimentRecipeSerializer,
 )
 from experimenter.experiments.tests.factories import ExperimentFactory
-from experimenter.projects.tests.factories import ProjectFactory
 
 
 class TestExperimentListView(TestCase):
@@ -30,34 +29,6 @@ class TestExperimentListView(TestCase):
 
         serialized_experiments = ExperimentSerializer(
             Experiment.objects.all(), many=True
-        ).data
-
-        self.assertEqual(serialized_experiments, json_data)
-
-    def test_list_view_filters_by_project_slug(self):
-        project = ProjectFactory.create()
-        project_experiments = []
-
-        # another projects experiments should be excluded
-        for i in range(2):
-            ExperimentFactory.create_with_variants()
-
-        # started project experiments should be included
-        for i in range(3):
-            experiment = ExperimentFactory.create_with_variants(
-                project=project
-            )
-            project_experiments.append(experiment)
-
-        response = self.client.get(
-            reverse("experiments-api-list"), {"project__slug": project.slug}
-        )
-        self.assertEqual(response.status_code, 200)
-
-        json_data = json.loads(response.content)
-
-        serialized_experiments = ExperimentSerializer(
-            project.experiments.all(), many=True
         ).data
 
         self.assertEqual(serialized_experiments, json_data)
