+++ app/pytest.ini
@@ -9,4 +9,3 @@ testpaths =
     experimenter/openidc/tests/
     experimenter/experiments/tests/
     experimenter/notifications/tests/
-    experimenter/projects/tests/
